Refactor code and add new tool for getting path of directory
